@@ -47,12 +47,8 @@ classifiers =
     License :: OSI Approved :: GNU General Public License v3 (GPLv3)
     Operating System :: POSIX :: Linux
     Programming Language :: Python
-    Programming Language :: Python :: 3.7
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
-    Programming Language :: Python :: 3.10
-    Programming Language :: Python :: 3.11
     Programming Language :: Python :: 3.12
+    Programming Language :: Python :: 3.13
     Programming Language :: Python :: 3 :: Only
     Programming Language :: Python :: Implementation :: CPython
     Topic :: Scientific/Engineering :: Atmospheric Science
@@ -60,10 +56,9 @@ classifiers =
 [options]
 packages = find_namespace:
 include_package_data = True
-python_requires = >=3.7
+python_requires = >=3.12
 install_requires =
     ansimarkup>=1.0.0
-    async-timeout>=3.0.0; python_version < "3.11"
     colorama>=0.4,<1
     graphql-core>=3.2,<3.3
     graphene>=3.4.0,<3.5
@@ -72,14 +67,11 @@ install_requires =
     metomi-isodatetime>=1!3.0.0,<1!3.2.0
     # Constrain protobuf version for compatible Scheduler-UIS comms across hosts
     packaging
-    protobuf>=4.24.4,<4.25.0
+    protobuf>=5,<8
     psutil>=5.6.0
     pyzmq>=22
-    importlib_metadata>=5.0; python_version < "3.12"
     # NOTE: exclude two urwid versions that were not compatible with Tui
     urwid>=2.2,!=2.6.2,!=2.6.3,<3
-    # unpinned transient dependencies used for type checking
-    tomli>=2; python_version < "3.11"
 
 [options.packages.find]
 include = cylc*
@@ -104,7 +96,7 @@ tests =
     aiosmtpd
     async_generator
     bandit>=1.7.0
-    coverage>=5.0.0,<7.3.1
+    coverage>=5.0.0
     flake8-broken-line>=0.3.0
     flake8-bugbear>=21.0.0
     flake8-builtins>=1.5.0
@@ -113,9 +105,9 @@ tests =
     flake8-implicit-str-concat>=0.4
     flake8-mutable>=1.2.0
     flake8-simplify>=0.14.0
-    flake8-type-checking; python_version > "3.7"
+    flake8-type-checking
     flake8>=3.0.0
-    mypy>=0.910,<1.9
+    mypy>=0.910
     # https://github.com/pytest-dev/pytest-asyncio/issues/706
     pytest-asyncio>=0.21.2,!=0.23.*
     pytest-cov>=2.8.0
@@ -123,13 +115,11 @@ tests =
     pytest-mock>=3.7
     pytest>=6
     testfixtures>=6.11.0
-    towncrier>=24.7.0; python_version > "3.7"
+    towncrier>=24.7.0
     # Type annotation stubs
     # http://mypy-lang.blogspot.com/2021/05/the-upcoming-switch-to-modular-typeshed.html
     types-Jinja2>=0.1.3
     types-protobuf>=0.1.10,!=5.29.1.20250402
-    types-six>=0.1.6
-    typing-extensions>=4
 tutorials =
     pillow
     requests
