@@ -29,15 +29,27 @@ from cylc.flow.commands import (
     set_prereqs_and_outputs,
 )
 from cylc.flow.cycling.integer import IntegerPoint
-from cylc.flow.data_messages_pb2 import PbTaskProxy
-from cylc.flow.data_store_mgr import TASK_PROXIES
+from cylc.flow.data_messages_pb2 import (
+    PbJob,
+    PbTaskProxy,
+)
+from cylc.flow.data_store_mgr import (
+    JOBS,
+    TASK_PROXIES,
+    task_mean_elapsed_time,
+)
+from cylc.flow.id import TaskTokens
 from cylc.flow.scheduler import Scheduler
 from cylc.flow.task_outputs import (
+    TASK_OUTPUT_FAILED,
     TASK_OUTPUT_STARTED,
     TASK_OUTPUT_SUBMITTED,
     TASK_OUTPUT_SUCCEEDED,
 )
+from cylc.flow.task_proxy import TaskProxy
 from cylc.flow.task_state import (
+    TASK_STATUS_FAILED,
+    TASK_STATUS_PREPARING,
     TASK_STATUS_SUCCEEDED,
     TASK_STATUS_WAITING,
 )
@@ -78,7 +90,10 @@ async def test_set_parentless_spawning(
     async with run(schd):
         # mark cycle 1 as succeeded
         schd.pool.set_prereqs_and_outputs(
-            ['1/a', '1/z'], ['succeeded'], [], ['1']
+            {TaskTokens('1', 'a'), TaskTokens('1', 'z')},
+            ['succeeded'],
+            [],
+            ['1'],
         )
 
         # the parentless task "a" should be spawned out to the runahead limit
@@ -108,7 +123,9 @@ async def test_rerun_incomplete(
     schd = scheduler(id_, paused_start=False)
     async with run(schd):
         # generate 1/a:x but do not complete 1/a
-        schd.pool.set_prereqs_and_outputs(['1/a'], ['x'], [], ['1'])
+        schd.pool.set_prereqs_and_outputs(
+            {TaskTokens('1', 'a')}, ['x'], [], ['1']
+        )
         triggers = reflog(schd)
         await complete(schd)
 
@@ -145,7 +162,7 @@ async def test_data_store(
 
         # set the 1/a:succeeded prereq of 1/z
         schd.pool.set_prereqs_and_outputs(
-            ['1/z'], [], ['1/a:succeeded'], ['1'])
+            {TaskTokens('1', 'z')}, [], ['1/a:succeeded'], ['1'])
         task_z = data[TASK_PROXIES][
             schd.pool.get_task(IntegerPoint('1'), 'z').tokens.id
         ]
@@ -153,14 +170,18 @@ async def test_data_store(
         assert task_z.prerequisites[0].satisfied is True
 
         # set 1/a:x the task should be waiting with output x satisfied
-        schd.pool.set_prereqs_and_outputs(['1/a'], ['x'], [], ['1'])
+        schd.pool.set_prereqs_and_outputs(
+            {TaskTokens('1', 'a')}, ['x'], [], ['1']
+        )
         await schd.update_data_structure()
         assert task_a.state == TASK_STATUS_WAITING
         assert task_a.outputs['x'].satisfied is True
         assert task_a.outputs['succeeded'].satisfied is False
 
         # set 1/a:succeeded the task should be succeeded with output x sat
-        schd.pool.set_prereqs_and_outputs(['1/a'], ['succeeded'], [], ['1'])
+        schd.pool.set_prereqs_and_outputs(
+            {TaskTokens('1', 'a')}, ['succeeded'], [], ['1']
+        )
         await schd.update_data_structure()
         assert task_a.state == TASK_STATUS_SUCCEEDED
         assert task_a.outputs['x'].satisfied is True
@@ -177,7 +198,9 @@ async def test_incomplete_detection(
     """It should detect and log finished tasks left with incomplete outputs."""
     schd = scheduler(flow(one_conf))
     async with start(schd):
-        schd.pool.set_prereqs_and_outputs(['1/one'], ['failed'], [], ['1'])
+        schd.pool.set_prereqs_and_outputs(
+            {TaskTokens('1', 'one')}, ['failed'], [], ['1']
+        )
     assert log_filter(contains='1/one did not complete')
 
 
@@ -188,7 +211,9 @@ async def test_pre_all(flow, scheduler, run):
     id_ = flow({'scheduling': {'graph': {'R1': 'a => z'}}})
     schd = scheduler(id_, paused_start=False)
     async with run(schd) as log:
-        schd.pool.set_prereqs_and_outputs(['1/z'], [], ['all'], [])
+        schd.pool.set_prereqs_and_outputs(
+            {TaskTokens('1', 'z')}, [], ['all'], []
+        )
         warn_or_higher = [i for i in log.records if i.levelno > 30]
         assert warn_or_higher == []
 
@@ -210,14 +235,18 @@ async def test_bad_prereq(
     })
     schd = scheduler(id_, paused_start=False)
     async with run(schd):
-        schd.pool.set_prereqs_and_outputs(['1/c'], [], ['1/a'], [])
+        schd.pool.set_prereqs_and_outputs(
+            {TaskTokens('1', 'c')}, [], ['1/a'], []
+        )
         assert schd.pool.get_task_ids() == {'1/a'}
         assert '1/c does not depend on "1/a:succeeded"' in caplog.text
 
         schd.workflow_db_mgr.process_queued_ops()
 
         # This will fail if the previous set left 1/c in the DB:
-        schd.pool.set_prereqs_and_outputs(['1/c'], [], ['1/b'], [])
+        schd.pool.set_prereqs_and_outputs(
+            {TaskTokens('1', 'c')}, [], ['1/b'], []
+        )
         assert schd.pool.get_task_ids() == {'1/a', '1/c'}
         await complete(schd, '1/c', timeout=5)
 
@@ -251,7 +280,7 @@ async def test_no_outputs_given(flow, scheduler, start):
     async with start(schd):
         foo = schd.pool.get_tasks()[0]
         await run_cmd(
-            set_prereqs_and_outputs(schd, [foo.identity], [])
+            set_prereqs_and_outputs(schd, {foo.tokens}, [])
         )
         assert set(foo.state.outputs.get_completed_outputs()) == {
             TASK_OUTPUT_SUBMITTED,
@@ -287,10 +316,147 @@ async def test_completion_expr(flow, scheduler, start):
     async with start(schd):
         foo = schd.pool.get_tasks()[0]
         await run_cmd(
-            set_prereqs_and_outputs(schd, [foo.identity], [])
+            set_prereqs_and_outputs(schd, {foo.tokens}, [])
         )
         assert set(foo.state.outputs.get_completed_outputs()) == {
             TASK_OUTPUT_SUBMITTED,
             TASK_OUTPUT_STARTED,
             TASK_OUTPUT_SUCCEEDED,
         }
+
+
+async def test_set_submitted(flow, scheduler, start):
+    """`cylc set --out submitted` should spawn children that depend on the
+    output, but not affect the task's state."""
+    schd: Scheduler = scheduler(flow('foo:submitted? => bar'))
+    async with start(schd):
+        foo = schd.pool.get_tasks()[0]
+        await run_cmd(
+            set_prereqs_and_outputs(
+                schd, [foo.identity], [], ['submitted']
+            )
+        )
+        assert set(foo.state.outputs.get_completed_outputs()) == {
+            TASK_OUTPUT_SUBMITTED,
+        }
+        assert schd.pool.get_task_ids() == {'1/foo', '1/bar'}
+        assert foo.state(TASK_STATUS_WAITING)
+
+
+async def test_job_state(flow, scheduler, start, db_select):
+    """Test that setting outputs does not change the job state."""
+    schd: Scheduler = scheduler(flow('foo => bar'))
+    fail_time = '1950-01-01T00:00:00Z'
+
+    def db_task_states(itask: TaskProxy):
+        return db_select(
+            schd, True, 'task_states', 'status', name=itask.tdef.name
+        )
+
+    def assert_job_failed(itask: TaskProxy):
+        """...in the DB and data store."""
+        assert db_select(
+            schd,
+            True,
+            'task_jobs',
+            'run_status',
+            'time_run_exit',
+            name=itask.tdef.name,
+        ) == [(1, fail_time)]
+        pb_job: PbJob = schd.data_store_mgr.data[schd.tokens.id][JOBS][
+            itask.job_tokens.id
+        ]
+        assert pb_job.state == TASK_STATUS_FAILED
+        assert pb_job.finished_time == fail_time
+
+    async with start(schd):
+        foo = schd.pool.get_tasks()[0]
+        schd.submit_task_jobs([foo])
+        schd.task_events_mgr.process_message(
+            foo, 'INFO', TASK_OUTPUT_FAILED, event_time=fail_time
+        )
+        await schd.update_data_structure()
+        assert foo.state(TASK_STATUS_FAILED)
+        assert db_task_states(foo) == [(TASK_STATUS_FAILED,)]
+        assert_job_failed(foo)
+
+        await run_cmd(
+            set_prereqs_and_outputs(schd, [foo.identity], [])
+        )
+        await schd.update_data_structure()
+        # Task is now succeeded:
+        assert foo.state(TASK_STATUS_SUCCEEDED)
+        assert db_task_states(foo) == [(TASK_STATUS_SUCCEEDED,)]
+        # But job is still failed:
+        assert_job_failed(foo)
+
+
+async def test_set_already_succeeded(
+    flow, scheduler, run, complete, db_select
+):
+    """Doing `cylc set` on a task that has already succeeded should not
+    change anything."""
+    schd: Scheduler = scheduler(
+        flow('foo => bar'),
+        paused_start=False,
+    )
+
+    def db_task_states(itask: TaskProxy):
+        return db_select(
+            schd,
+            True,
+            'task_states',
+            'submit_num',
+            'status',
+            'time_updated',
+            name=itask.tdef.name,
+        )
+
+    def data_store_task_state(itask: TaskProxy):
+        return schd.data_store_mgr.data[schd.tokens.id][TASK_PROXIES][
+            itask.tokens.id
+        ].state
+
+    async with run(schd):
+        foo = schd.pool.get_tasks()[0]
+        await complete(schd, foo.identity)
+        time_updated = db_task_states(foo)[0][2]
+        expected = [(1, TASK_STATUS_SUCCEEDED, time_updated)]
+        assert db_task_states(foo) == expected
+        assert data_store_task_state(foo) == TASK_STATUS_SUCCEEDED
+
+        await run_cmd(
+            set_prereqs_and_outputs(schd, [foo.identity], [])
+        )
+        assert foo.state(TASK_STATUS_SUCCEEDED)
+        await schd.update_data_structure()
+        assert db_task_states(foo) == expected
+        assert data_store_task_state(foo) == TASK_STATUS_SUCCEEDED
+
+
+async def test_timings(one_conf, flow, scheduler, start, caplog):
+    """Test that setting outputs does not change the task timings."""
+    wid = flow({
+        **one_conf,
+        'runtime': {
+            'one': {
+                'execution time limit': 'PT100S',
+            },
+        },
+    })
+    schd: Scheduler = scheduler(wid)
+    async with start(schd):
+        itask = schd.pool.get_tasks()[0]
+
+        def check_times():
+            assert not itask.tdef.elapsed_times
+            assert task_mean_elapsed_time(itask.tdef) == 100
+
+        itask.state_reset(TASK_STATUS_PREPARING)
+        schd.task_events_mgr.process_message(
+            itask, 'INFO', TASK_OUTPUT_STARTED
+        )
+        check_times()
+
+        await run_cmd(set_prereqs_and_outputs(schd, [itask.identity], []))
+        check_times()
