@@ -32,6 +32,7 @@ from cylc.flow.commands import (
 )
 from cylc.flow.cycling.integer import IntegerPoint
 from cylc.flow.cycling.iso8601 import ISO8601Point
+from cylc.flow.id import TaskTokens, Tokens
 from cylc.flow.network.resolvers import TaskMsg
 from cylc.flow.task_events_mgr import (
     TaskEventsManager,
@@ -148,7 +149,7 @@ async def test_task_completion(
         }:
             # set the combination of outputs
             schd.pool.set_prereqs_and_outputs(
-                ['1/a'],
+                {TaskTokens('1', 'a')},
                 combination,
                 [],
                 ['1'],
@@ -164,7 +165,7 @@ async def test_task_completion(
         for combination in completion_outputs:
             # set the combination of outputs
             schd.pool.set_prereqs_and_outputs(
-                ['1/a'],
+                {TaskTokens('1', 'a')},
                 combination,
                 [],
                 ['1'],
@@ -207,7 +208,7 @@ async def test_expire_orthogonality(flow, scheduler, start):
         # tell the scheduler that the task *submit-failed*
         schd.message_queue.put(
             TaskMsg(
-                '1/a/01',
+                Tokens('//1/a/01'),
                 '2000-01-01T00:00:00+00',
                 'INFO',
                 TaskEventsManager.EVENT_SUBMIT_FAILED
@@ -221,7 +222,7 @@ async def test_expire_orthogonality(flow, scheduler, start):
         # tell the scheduler that the task *failed*
         schd.message_queue.put(
             TaskMsg(
-                '1/a/01',
+                Tokens('//1/a/01'),
                 '2000-01-01T00:00:00+00',
                 'INFO',
                 TaskEventsManager.EVENT_FAILED,
@@ -235,7 +236,7 @@ async def test_expire_orthogonality(flow, scheduler, start):
         # tell the scheduler that the task *expired*
         schd.message_queue.put(
             TaskMsg(
-                '1/a/01',
+                Tokens('//1/a/01'),
                 '2000-01-01T00:00:00+00',
                 'INFO',
                 TaskEventsManager.EVENT_EXPIRED,
